[BugFix] AutonomousPublishTriggerTest: savedVersionDelta is int not long

+++ fe/fe-core/src/test/java/com/starrocks/lake/compaction/AutonomousPublishTriggerTest.java
@@ -30,7 +30,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  */
 public class AutonomousPublishTriggerTest {
 
-    private long savedVersionDelta;
+    private int savedVersionDelta;
     private double savedHighScore;
     private int savedMinDeltaForHighScore;
     private long savedMaxIntervalMs;
